fix(audit): restore performance hotspot budgets
