+++ src/options/app/productionStitchStateMapper.ts
@@ -1,8 +1,7 @@
 import { normalizeUsageStats } from '@shared/constants';
 import { previewContent } from '@options/stitch/content';
 import { prepareUsageHistory } from '@options/stitch/usageHistory';
-import { getService, hasService, TOKENS } from '@shared/di';
-import type { PlatformServices } from '@platform/types';
+import { resolveExtensionVersionLabel } from './productionStitchVersion';
 import type { CompleteOptions, InterfaceTheme, StoredOptions } from '@shared/types/options';
 import type { UsageStats } from '@shared/types/usage';
 import type { YamlConfigOverrides } from '@shared/types/yamlConfig';
@@ -12,6 +11,8 @@ import type {
   RoutingRule,
   VaultRecord
 } from '@options/stitch/types';
+
+export { resolveExtensionVersionLabel } from './productionStitchVersion';
 
 const MODIFIER_LABEL_TO_OPTION = {
   Alt: 'alt',
@@ -29,7 +30,6 @@ const MODIFIER_OPTION_TO_LABEL = {
 
 export const RUNTIME_SURFACE_RESOURCE_IDS = new Set(['clipper', 'reader', 'video', 'task-success']);
 export const LEGACY_USAGE_STATS_STORAGE_KEY = 'usage_stats';
-const PACKAGE_VERSION = '0.2.0';
 export const HIGHLIGHT_THEME_CLASSES: Record<
   CompleteOptions['readingSession']['highlightTheme'],
   string
@@ -361,29 +361,6 @@ export function toTemplateValues(options: CompleteOptions): Record<string, strin
     readingCustom: options.templates.reading,
     aiChat: options.templates.ai
   };
-}
-
-type ExtensionVersionReader = () => string | undefined;
-
-function readPlatformManifestVersion(): string | undefined {
-  if (!hasService(TOKENS.platformServices)) {
-    return undefined;
-  }
-  return getService<PlatformServices>(TOKENS.platformServices).runtime.getManifest?.()?.version;
-}
-
-export function resolveExtensionVersionLabel(
-  readVersion: ExtensionVersionReader = readPlatformManifestVersion
-): string {
-  try {
-    const version = readVersion();
-    if (typeof version === 'string' && version.trim().length > 0) {
-      return `v${version.trim()}`;
-    }
-  } catch {
-    // Platform services can be unavailable in isolated preview or unit-test contexts.
-  }
-  return `v${PACKAGE_VERSION}`;
 }
 
 function resolveUsageStatsFromOptions(options: CompleteOptions): UsageStats {
